Move a couple functions out of main.py
